+++ test/cant_wait_test.rb
@@ -6,11 +6,11 @@ $LOAD_PATH.unshift(File.dirname(__FILE__))  # We require the rails environment f
 
 class CantWait1Test < MiniTest::Unit::TestCase
 
-  TEST_RAILS_APP = [{ version: '3.0.3',     rails_root: 'test_apps/Test_3_0_3',    },
+  TEST_RAILS_APP = [{ version: '3.0.3',     rails_root: 'test_apps/Test_3_0_03',    },
                     { version: '3.0.20',    rails_root: 'test_apps/Test_3_0_20',   },
                     { version: '3.1.12',    rails_root: 'test_apps/Test_3_1_12',   },
                     { version: '3.2.13',    rails_root: 'test_apps/Test_3_2_13'    },
-                    { version: '4.0.0.rc1', rails_root: 'test_apps/Test_4_0_0_rc1' }]
+                    { version: '4.0.0.rc2', rails_root: 'test_apps/Test_4_0_0_rc2' }]
 
   # Set a random timeout and rails app to be tested
   def setup