+++ test/test_apps/Test_4_0_0_rc2/config/environments/test.rb
@@ -1,4 +1,4 @@
-Test400Rc1::Application.configure do
+Test400Rc2::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
   # The test environment is used exclusively to run your application's