+++ test/test_apps/Test_4_0_0_rc2/config/initializers/session_store.rb
@@ -1,0 +1,3 @@
+# Be sure to restart your server when you modify this file.
+
+Test400Rc2::Application.config.session_store :cookie_store, key: '_Test_4_0_0_rc2_session'